requirements.txt
chainlit==1.0.0
google-cloud-dialogflow-cx
python-dotenv==1.0.0